+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Current (in progress)
 
-- Nothing yet
+- Add migration to roolback on resource's schema's name to None [#2615](https://github.com/opendatateam/udata/pull/2615)
 
 ## 2.7.0 (2021-05-25)